+++ packages/analyzer/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @cognidocs/analyzer
 
+## 1.0.2
+
+### Patch Changes
+
+- Add cognidocs wrapper for all packages
+- Updated dependencies
+  - @cognidocs/types@1.0.1
+  - @cognidocs/utils@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes