+++ packages/graph-viz/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @cognidocs/graph-viz
 
+## 1.0.2
+
+### Patch Changes
+
+- Add cognidocs wrapper for all packages
+
 ## 1.0.1
 
 ### Patch Changes