+++ packages/site-builder/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @cognidocs/site-builder
 
+## 1.0.2
+
+### Patch Changes
+
+- Add cognidocs wrapper for all packages
+- Updated dependencies
+  - @cognidocs/graph-viz@1.0.2
+  - @cognidocs/utils@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes